start using dispatch selectors from the "coopcycle-frontend-js" library
