+++ js/app/dashboard/components/Task.js
@@ -6,7 +6,9 @@ import { ContextMenuTrigger } from 'react-contextmenu'
 import _ from 'lodash'
 
 import { setCurrentTask, toggleTask, selectTask } from '../redux/actions'
-import { selectTasksWithColor, selectIsVisibleTask } from '../redux/selectors'
+import { selectIsVisibleTask } from '../redux/selectors'
+import { selectSelectedDate, selectTasksWithColor } from 'coopcycle-frontend-js/dispatch/redux'
+
 import { addressAsText } from '../utils'
 import TaskEta from './TaskEta'
 
@@ -201,11 +203,11 @@ function mapStateToProps(state, ownProps) {
   return {
     selected: -1 !== state.selectedTasks.indexOf(ownProps.task),
     color,
-    date: state.date,
+    date: selectSelectedDate(state),
     isVisible: selectIsVisibleTask({
       task: ownProps.task,
       filters: state.filters,
-      date: state.date,
+      date: selectSelectedDate(state),
     }),
   }
 }
