+++ js/app/dashboard/redux/reducers.js
@@ -1,22 +1,14 @@
 import _ from 'lodash'
-import Moment from 'moment'
-import { extendMoment } from 'moment-range'
+import { moment } from 'coopcycle-frontend-js'
 
-import { createTaskList, removedTasks, withoutTasks } from './utils'
 import {
-  UPDATE_TASK,
   OPEN_ADD_USER,
   CLOSE_ADD_USER,
-  MODIFY_TASK_LIST_REQUEST,
-  MODIFY_TASK_LIST_REQUEST_SUCCESS,
-  TASK_LIST_UPDATED,
   TOGGLE_POLYLINE,
   TOGGLE_TASK,
   SELECT_TASK,
   SELECT_TASKS,
   SET_TASK_LIST_GROUP_MODE,
-  ADD_TASK_LIST_REQUEST,
-  ADD_TASK_LIST_REQUEST_SUCCESS,
   SET_GEOLOCATION,
   SET_OFFLINE,
   OPEN_NEW_TASK_MODAL,
@@ -41,7 +33,6 @@ import {
   LOAD_TASK_EVENTS_REQUEST,
   LOAD_TASK_EVENTS_SUCCESS,
   LOAD_TASK_EVENTS_FAILURE,
-  SET_TASK_LISTS_LOADING,
   ADD_IMPORT,
   IMPORT_SUCCESS,
   IMPORT_ERROR,
@@ -50,40 +41,6 @@ import {
   SET_CLUSTERS_ENABLED,
   CLEAR_SELECTED_TASKS,
 } from './actions'
-
-const moment = extendMoment(Moment)
-
-const replaceOrAddTask = (tasks, task) => {
-
-  const taskIndex = _.findIndex(tasks, t => t['@id'] === task['@id'])
-
-  if (-1 !== taskIndex) {
-
-    const newTasks = tasks.slice(0)
-    newTasks.splice(taskIndex, 1, Object.assign({}, tasks[taskIndex], task))
-
-    return newTasks
-  }
-
-  return tasks.concat([ task ])
-}
-
-const removeTask = (tasks, task) => _.filter(tasks, t => t['@id'] !== task['@id'])
-
-const acceptTask = (task, date) => {
-
-  const dateAsRange = moment.range(
-    moment(date).set({ hour:  0, minute:  0, second:  0 }),
-    moment(date).set({ hour: 23, minute: 59, second: 59 })
-  )
-
-  const range = moment.range(
-    moment(task.doneAfter),
-    moment(task.doneBefore)
-  )
-
-  return range.overlaps(dateAsRange)
-}
 
 const defaultFilters = {
   showFinishedTasks: true,
@@ -95,13 +52,6 @@ const defaultFilters = {
 }
 
 const initialState = {
-
-  allTasks: [],
-  unassignedTasks: [],
-  taskLists: [],
-  date: moment(),
-
-  taskListsLoading: false,
   addModalIsOpen: false,
   polylineEnabled: {},
   taskListGroupMode: 'GROUP_MODE_FOLDERS',
@@ -123,7 +73,6 @@ const initialState = {
   settingsModalIsOpen: false,
   polylineStyle: 'normal',
   searchIsOn: false,
-  tasksWithColor: {},
   isLoadingTaskEvents: false,
   taskEvents: {},
   imports: {},
@@ -133,161 +82,12 @@ const initialState = {
   clustersEnabled: false,
 }
 
-const rootReducer = (state = initialState, action) => {
-  let newTaskLists = state.taskLists.slice(0)
-  let taskListIndex
-
-  switch (action.type) {
-  case MODIFY_TASK_LIST_REQUEST:
-
-    taskListIndex = _.findIndex(state.taskLists, taskList => taskList.username === action.username)
-    newTaskLists.splice(taskListIndex, 1, {
-      ...state.taskLists[taskListIndex],
-      items: action.tasks,
-    })
-
-    let removed = removedTasks(state.taskLists[taskListIndex].items, action.tasks)
-
-    return {
-      ...state,
-      taskListsLoading: true,
-      taskLists: newTaskLists,
-      unassignedTasks: withoutTasks(
-        Array.prototype.concat(state.unassignedTasks, removed),
-        action.tasks
-      ),
-    }
-
-  case MODIFY_TASK_LIST_REQUEST_SUCCESS:
-
-    taskListIndex = _.findIndex(state.taskLists, taskList => taskList['@id'] === action.taskList['@id'])
-    newTaskLists.splice(taskListIndex, 1, {
-      ...action.taskList,
-      items: action.taskList.items,
-    })
-
-    return {
-      ...state,
-      taskLists: newTaskLists,
-    }
-
-  case ADD_TASK_LIST_REQUEST_SUCCESS:
-
-    return {
-      ...state,
-      taskLists: Array.prototype.concat(state.taskLists, action.taskList),
-    }
-
-  case UPDATE_TASK:
-
-    if (!acceptTask(action.task, state.date)) {
-      return state
-    }
-
-    let newUnassignedTasks = state.unassignedTasks.slice(0)
-    let unassignedTasksIndex = _.findIndex(state.unassignedTasks, task => task['@id'] === action.task['@id'])
-    let taskListsIndex = _.findIndex(state.taskLists, taskList => {
-      return _.includes(_.map(taskList.items, task => task['@id']), action.task['@id'])
-    })
-
-    if (-1 !== unassignedTasksIndex) {
-      if (action.task.isAssigned) {
-        newUnassignedTasks = removeTask(state.unassignedTasks, action.task)
-      } else {
-        newUnassignedTasks = replaceOrAddTask(state.unassignedTasks, action.task)
-      }
-    } else {
-      if (!action.task.isAssigned) {
-        newUnassignedTasks = replaceOrAddTask(state.unassignedTasks, action.task)
-      }
-    }
-
-    if (action.task.isAssigned) {
-
-      let targetTaskListsIndex = _.findIndex(state.taskLists, taskList => taskList.username === action.task.assignedTo)
-
-      if (-1 !== taskListsIndex) {
-        if (targetTaskListsIndex !== taskListsIndex) {
-          newTaskLists.splice(taskListsIndex, 1, {
-            ...state.taskLists[taskListsIndex],
-            items: removeTask(state.taskLists[taskListsIndex].items, action.task)
-          })
-        }
-      }
-
-      if (-1 !== targetTaskListsIndex) {
-        newTaskLists.splice(targetTaskListsIndex, 1, {
-          ...state.taskLists[targetTaskListsIndex],
-          items: replaceOrAddTask(state.taskLists[targetTaskListsIndex].items, action.task)
-        })
-      } else {
-        newTaskLists.push(
-          createTaskList(action.task.assignedTo, [ action.task ])
-        )
-      }
-
-    } else {
-      if (-1 !== taskListsIndex) {
-        newTaskLists.splice(taskListsIndex, 1, {
-          ...state.taskLists[taskListsIndex],
-          items: removeTask(state.taskLists[taskListsIndex].items, action.task)
-        })
-      }
-    }
-
-    return {
-      ...state,
-      unassignedTasks: newUnassignedTasks,
-      taskLists: newTaskLists,
-      allTasks: _.uniqBy(Array.prototype.concat(state.allTasks, [ action.task ]), '@id'),
-    }
-
-  case TASK_LIST_UPDATED:
-
-    taskListIndex = _.findIndex(state.taskLists, taskList => taskList['@id'] === action.taskList['@id'])
-
-    if (-1 === taskListIndex) {
-
-      return state
-    }
-
-    newTaskLists.splice(taskListIndex, 1, {
-      ...state.taskLists[taskListIndex],
-      distance: action.taskList.distance,
-      duration: action.taskList.duration,
-      polyline: action.taskList.polyline,
-    })
-
-    return {
-      ...state,
-      taskLists: newTaskLists,
-    }
-
-  }
-
-  return state
-}
-
 const addModalIsOpen = (state = false, action) => {
   switch(action.type) {
   case OPEN_ADD_USER:
     return true
   case CLOSE_ADD_USER:
     return false
-  default:
-    return state
-  }
-}
-
-const _taskListsLoading = (state = false, action) => {
-  switch(action.type) {
-  case ADD_TASK_LIST_REQUEST:
-    return true
-  case ADD_TASK_LIST_REQUEST_SUCCESS:
-  case MODIFY_TASK_LIST_REQUEST_SUCCESS:
-    return false
-  case SET_TASK_LISTS_LOADING:
-    return action.loading
   default:
     return state
   }
@@ -421,19 +221,6 @@ const offline = (state = [], action) => {
   }
 
   return state
-}
-
-const combinedTasks = (state = initialState, action) => {
-
-  const { unassignedTasks, taskLists, allTasks, taskListsLoading } = rootReducer(state, action)
-
-  return {
-    ...state,
-    unassignedTasks,
-    taskLists,
-    allTasks,
-    taskListsLoading,
-  }
 }
 
 const taskModalIsOpen = (state = false, action) => {
@@ -668,15 +455,10 @@ const clustersEnabled = (state = initialState.clustersEnabled, action) => {
 
 export default (state = initialState, action) => {
 
-  const { allTasks, unassignedTasks, taskLists, tasksWithColor, taskListsLoading } = combinedTasks(state, action)
   const { filters, isDefaultFilters } = combinedFilters(state, action)
 
   return {
     ...state,
-    unassignedTasks,
-    taskLists,
-    allTasks,
-    taskListsLoading: _taskListsLoading(taskListsLoading, action),
     addModalIsOpen: addModalIsOpen(state.addModalIsOpen, action),
     polylineEnabled: polylineEnabled(state.polylineEnabled, action),
     taskListGroupMode: taskListGroupMode(state.taskListGroupMode, action),
@@ -694,7 +476,6 @@ export default (state = initialState, action) => {
     searchIsOn: searchIsOn(state.searchIsOn, action),
     settingsModalIsOpen: settingsModalIsOpen(state.settingsModalIsOpen, action),
     polylineStyle: polylineStyle(state.polylineStyle, action),
-    tasksWithColor,
     isLoadingTaskEvents: isLoadingTaskEvents(state.isLoadingTaskEvents, action),
     taskEvents: taskEvents(state.taskEvents, action),
     imports: imports(state.imports, action),
