+++ js/app/dashboard/redux/selectors.js
@@ -1,25 +1,7 @@
 import { createSelector } from 'reselect'
-import Moment from 'moment'
-import { extendMoment } from 'moment-range'
-import { filter, forEach, includes, intersectionWith, isEqual } from 'lodash'
-import { mapToColor } from 'coopcycle-frontend-js'
-
-const moment = extendMoment(Moment)
-
-function flattenTaskLists(taskLists) {
-  const tasks = []
-  forEach(taskLists, taskList => taskList.items.forEach(task => tasks.push(task)))
-
-  return tasks
-}
-
-export const selectTasks = createSelector(
-  state => state.unassignedTasks,
-  state => flattenTaskLists(state.taskLists),
-  (unassignedTasks, assignedTasks) => {
-    return unassignedTasks.slice(0).concat(assignedTasks)
-  }
-)
+import { moment } from 'coopcycle-frontend-js'
+import { selectTaskLists } from 'coopcycle-frontend-js/dispatch/redux'
+import { filter, includes, intersectionWith, isEqual } from 'lodash'
 
 export const selectFilteredTasks = createSelector(
   state => state.tasks,
@@ -39,7 +21,7 @@ export const selectFilteredTasks = createSelector(
 )
 
 export const selectBookedUsernames = createSelector(
-  state => state.taskLists,
+  selectTaskLists,
   taskLists => taskLists.map(taskList => taskList.username)
 )
 
@@ -123,9 +105,4 @@ export const selectIsVisibleTask = createSelector(
 
     return true
   }
-)
-
-export const selectTasksWithColor = createSelector(
-  state => state.allTasks,
-  allTasks => mapToColor(allTasks)
 )
