+++ js/app/dashboard/redux/store.js
@@ -1,6 +1,9 @@
 import { createStore, applyMiddleware, compose } from 'redux'
 import thunk from 'redux-thunk'
-import reducer from './reducers'
+import { reducer as coreReducer } from 'coopcycle-frontend-js/dispatch/redux'
+import webReducer from './reducers'
+import webDispatchReducer from './dispatchReducers'
+import reduceReducers from 'reduce-reducers';
 import { socketIO, persistFilters } from './middlewares'
 
 const middlewares = [ thunk, socketIO, persistFilters ]
@@ -11,8 +14,17 @@ const middlewares = [ thunk, socketIO, persistFilters ]
 const composeEnhancers = (typeof window !== 'undefined' &&
   window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) || compose
 
-export const createStoreFromPreloadedState = preloadedState => {
+const reducer = (state, action) => {
+  const rootState = webReducer(state, action)
+  const dispatchState = reduceReducers(coreReducer, webDispatchReducer)(state.dispatch, action)
 
+  return {
+    ...rootState,
+    dispatch: dispatchState,
+  }
+}
+
+export const createStoreFromPreloadedState = preloadedState => {
   return createStore(
     reducer,
     preloadedState,
